src/requirements.txt
# LangGraph & LangChain Core
langgraph>=0.2.70
langchain>=0.3.9
langchain-core>=0.3.75
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langsmith>=0.2.0
langfuse>=2.55.0

# LangGraph CLI & API (Required for development server)
# Note: May need compatibility updates for v1-alpha
langgraph-cli[inmem]>=0.4.0
langgraph-api>=0.4.7

# LangGraph Supervisor Dependencies (compatible with v1-alpha)
langgraph-supervisor>=0.0.29

# Agent & UI Framework
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database & Memory
aiosqlite>=0.20.0
redis>=5.2.0
langgraph-checkpoint-sqlite>=2.0.0
langgraph-checkpoint-postgres>=2.0.0
supabase>=2.10.0

# MCP (Model Context Protocol) with LangChain adapters
mcp>=1.9.2
langchain-mcp-adapters>=0.1.9

# Google APIs & OAuth
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.150.0

# HTTP & Async
httpx>=0.28.0
aiohttp>=3.11.0
requests>=2.32.0

# Authentication & JWT
pyjwt[crypto]>=2.8.0
cryptography>=42.0.5

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.9.0
typer>=0.15.0
PyYAML>=6.0.2
psutil>=6.1.0
pytz>=2024.1

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
ruff>=0.9.4
mypy>=1.8.0
black>=24.0.0

# Optional AI Models
openai>=1.55.0
anthropic>=0.39.0

# Job Search Agent Dependencies
pypdf>=5.0.0
html2text>=2024.2.26
beautifulsoup4>=4.12.0
md2pdf>=1.0.0
pandas>=2.2.0
selenium>=4.25.0
webdriver-manager>=4.0.0
fastmcp>=0.5.0
